pyproject.toml
[tool.poetry]
name = "blog_api_project"
version = "0.1.0"
description = ""
authors = ["Leen Hazaimeh <leenhazaimeh94@gmail.coom>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
